New test program for using widgets
This is kind of a proof-of-concept sandbox, to see if I could write a
program that uses widgets to control various annotations for collidl in
a single window.  I think the answer is yes, this is doable.  The next
step is to port many of these ideas and helper functions into the
collidle program.