Major rewrite: interactive graphics using widgets
This is a major rewrite of much of the code.  I've put several graphics
together in one window using the "new" (version 8.0) IDL graphics, and
also used radio buttons and other things to select different annotations
to show, resize the windows, and so on.  With this revision, all of the
"old" graphics (pre 8.0) are deprecated.